+++ app/src/main/AndroidManifest.xml
@@ -77,7 +77,7 @@
         </activity>
 
         <activity
-            android:name=".Main.AddEditLabelActivity"
+            android:name=".AddEditLabels.AddEditLabelActivity"
             android:label="@string/label_dialog_edit"
             android:parentActivityName=".Main.TimerActivity">
         </activity>
